internal/shader: bug fix: huge array lengths accepted
An array type had no upper bound on its length. The backends emit code
for each element, so a declaration such as var a [20000000]float made
code generation take an enormous amount of time and memory.

Reject an array longer than 4096 elements. A uniform variable cannot be
longer than that on DirectX 11 or Metal anyway. An existing test case
that expected a [2147483647]int array to compile now uses the largest
allowed length.

+++ internal/shader/expr.go
@@ -1112,7 +1112,7 @@ func (cs *compileState) parseExpr(block *block, fname string, expr ast.Expr, mar
 		var ok bool
 		if at, isArray := e.Type.(*ast.ArrayType); isArray && isEllipsis(at.Len) {
 			// The length of a [...] array is the number of the elements.
-			t, ok = cs.parseArrayType(block, fname, at, len(e.Elts))
+			t, ok = cs.parseArrayType(block, fname, at, int64(len(e.Elts)))
 		} else {
 			t, ok = cs.parseType(block, fname, e.Type)
 		}
